+++ TablePro/Theme/ResolvedThemeColors.swift
@@ -16,6 +16,8 @@ struct ResolvedEditorColors {
     let lineNumberSwiftUI: Color
     let invisibles: NSColor
     let invisiblesSwiftUI: Color
+    let currentStatementHighlight: NSColor
+    let currentStatementHighlightSwiftUI: Color
 
     let keyword: NSColor
     let keywordSwiftUI: Color
@@ -49,6 +51,8 @@ struct ResolvedEditorColors {
         lineNumberSwiftUI = colors.lineNumber.swiftUIColor
         invisibles = colors.invisibles.nsColor
         invisiblesSwiftUI = colors.invisibles.swiftUIColor
+        currentStatementHighlight = colors.currentStatementHighlight.nsColor
+        currentStatementHighlightSwiftUI = colors.currentStatementHighlight.swiftUIColor
 
         keyword = colors.syntax.keyword.nsColor
         keywordSwiftUI = colors.syntax.keyword.swiftUIColor